Add config module to look for credentials
The config.py module checks that the Spotify Client ID and the
Client Secrete credentials are present in the .env file. These
credentials are used to obtain an access token to the Spotify Web API.
Simple error-handling is used to ensure that these credentials are
present, but the module currently does not ensure that the credentials
are valid.